my-collection-api/requirements.txt
flask
marshmallow
sqlalchemy
flasgger
flask-cor
